ver_17

+++ admin/includes/view_all_users.php
@@ -45,7 +45,7 @@
         <?php
             if(isset($_GET['change_to_admin']))
             {
-                $the_change_to_admin_id = $_GET['change_to_admin'];
+                $the_change_to_admin_id = escape($_GET['change_to_admin']);
 
                 $query = "UPDATE users SET user_role = 'Admin' WHERE user_id = {$the_change_to_admin_id}";
 
@@ -59,7 +59,7 @@
 
             if(isset($_GET['change_to_subscriber']))
             {
-                $the_change_to_sub_id = $_GET['change_to_subscriber'];
+                $the_change_to_sub_id = escape($_GET['change_to_subscriber']);
 
                 $query = "UPDATE users SET user_role = 'Subscriber' WHERE user_id = {$the_change_to_sub_id}";
 
@@ -73,15 +73,21 @@
 
             if(isset($_GET['delete']))
             {
-                $the_user_id = $_GET['delete'];
+                if(isset($_SESSION['user_role']))
+                {
+                    if($_SESSION['user_role'] === 'admin')
+                    {
+                        $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
 
-                $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+                        $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
 
-                $delete_user_query = mysqli_query($connection, $query);
+                        $delete_user_query = mysqli_query($connection, $query);
 
-                comfirmQuery($delete_user_query);
+                        comfirmQuery($delete_user_query);
 
-                header("Location: users.php");
+                        header("Location: users.php");
+                    }
+                }
             }
         ?>
     </tbody>
